+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.31.0] - 2023-02-04
+
 - configure logging through `log-level` / `CMEEL_LOG_LEVEL`
 - DEBUG level: show pip freeze, the commands, and their environment
 
@@ -236,6 +238,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial minial working example
 
 [Unreleased]: https://github.com/cmake-wheel/cmeel/compare/v0.4.2...main
+[v0.31.0]: https://github.com/cmake-wheel/cmeel/compare/v0.30.0...v0.31.0
 [v0.30.0]: https://github.com/cmake-wheel/cmeel/compare/v0.29.0...v0.30.0
 [v0.29.0]: https://github.com/cmake-wheel/cmeel/compare/v0.28.0...v0.29.0
 [v0.28.0]: https://github.com/cmake-wheel/cmeel/compare/v0.27.0...v0.28.0